requirements.txt
streamlit
pytesseract
pdf2image
Pillow
pdfplumber
opencv-python

google-genai
python-dotenv
python-dateutil
